Refactor: Create IconTextInput component for search bar
