+++ packages/shared/components/inputs/index.js
@@ -18,4 +18,5 @@ export { default as RecipientInput } from './RecipientInput.svelte'
 export { default as TextInput } from './TextInput.svelte'
 export { default as Toggle } from './Toggle.svelte'
 export { default as TogglableButton } from './TogglableButton.svelte'
+export { default as IconTextInput } from './IconTextInput.svelte'
 export { default as UnitInput } from './UnitInput.svelte'
