+++ packages/shared/components/molecules/index.js
@@ -1,6 +1,5 @@
 export { default as AssetList } from './AssetList.svelte'
 export { default as ExpirationTimePicker } from './ExpirationTimePicker.svelte'
 export { default as TransactionDetails } from './TransactionDetails.svelte'
-export { default as SearchBar } from './SearchBar.svelte'
 export { default as RewardClaimList } from './RewardClaimList.svelte'
 export * from './menu-items'
