+++ packages/shared/components/organisms/index.js
@@ -1,3 +1,3 @@
 export { default as AccountAssetsList } from './AccountAssetsList.svelte'
-export { default as AccountHistory } from './AccountHistory.svelte'
+export { default as AccountActivity } from './AccountActivity.svelte'
 export { default as AccountSummary } from './AccountSummary.svelte'
