--- packages/shared/components/panes/wallet/AccountHistoryPane.svelte
@@ -1,7 +1,0 @@
-<script lang="typescript">
-    import { Pane, AccountHistory } from 'shared/components'
-</script>
-
-<Pane>
-    <AccountHistory />
-</Pane>
