+++ packages/shared/components/panes/wallet/index.js
@@ -1,4 +1,4 @@
-export { default as AccountHistoryPane } from './AccountHistoryPane.svelte'
+export { default as AccountActivityPane } from './AccountActivityPane.svelte'
 export { default as AccountSummaryAndAssetsPane } from './AccountSummaryAndAssetsPane.svelte'
 export { default as AssetsPane } from './AssetsPane.svelte'
 export { default as BarChartPane } from './BarChartPane.svelte'
